refactor: add reusuable ResultDetail component and refactor code
